// src/shared/api/generated/model/getAnalyticsDashboardCardsParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */
import type { GetAnalyticsDashboardCardsPlatformsItem } from './getAnalyticsDashboardCardsPlatformsItem';
import type { GetAnalyticsDashboardCardsStatus } from './getAnalyticsDashboardCardsStatus';

export type GetAnalyticsDashboardCardsParams = {
'campaign_ids[]'?: string[];
'creator_ids[]'?: string[];
'platforms[]'?: GetAnalyticsDashboardCardsPlatformsItem[];
'countries[]'?: string[];
status?: GetAnalyticsDashboardCardsStatus;
range_preset?: string;
range_start?: string;
range_end?: string;
};