// src/shared/api/generated/model/getAnalyticsDashboardTopCreatorsParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */
import type { GetAnalyticsDashboardTopCreatorsPlatformsItem } from './getAnalyticsDashboardTopCreatorsPlatformsItem';
import type { GetAnalyticsDashboardTopCreatorsSortBy } from './getAnalyticsDashboardTopCreatorsSortBy';
import type { GetAnalyticsDashboardTopCreatorsStatus } from './getAnalyticsDashboardTopCreatorsStatus';

export type GetAnalyticsDashboardTopCreatorsParams = {
'campaign_ids[]'?: string[];
'creator_ids[]'?: string[];
'platforms[]'?: GetAnalyticsDashboardTopCreatorsPlatformsItem[];
'countries[]'?: string[];
status?: GetAnalyticsDashboardTopCreatorsStatus;
range_preset?: string;
range_start?: string;
range_end?: string;
sort_by?: GetAnalyticsDashboardTopCreatorsSortBy;
};