// src/shared/api/generated/model/getAnalyticsDashboardTopVideosParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */
import type { GetAnalyticsDashboardTopVideosPlatformsItem } from './getAnalyticsDashboardTopVideosPlatformsItem';
import type { GetAnalyticsDashboardTopVideosSortBy } from './getAnalyticsDashboardTopVideosSortBy';
import type { GetAnalyticsDashboardTopVideosStatus } from './getAnalyticsDashboardTopVideosStatus';

export type GetAnalyticsDashboardTopVideosParams = {
'campaign_ids[]'?: string[];
'creator_ids[]'?: string[];
'platforms[]'?: GetAnalyticsDashboardTopVideosPlatformsItem[];
'countries[]'?: string[];
status?: GetAnalyticsDashboardTopVideosStatus;
range_preset?: string;
range_start?: string;
range_end?: string;
sort_by?: GetAnalyticsDashboardTopVideosSortBy;
};